# Avoid logging in common case.