--- a/impl/build.gradle
+++ b/impl/build.gradle
@@ -1,3 +1,7 @@
+plugins {
+    id "me.champeau.jmh"
+}
+
 description = "PerfMark Implementation API"
 ext.moduleName =  "io.perfmark.impl"
 ext.jdkVersion = JavaVersion.VERSION_1_6
@@ -21,3 +25,39 @@ dependencies {
             libraries.errorprone
 }
 
+
+jmh {
+
+    timeOnIteration = "1s"
+    warmup = "1s"
+    fork = 400
+    warmupIterations = 0
+
+    includes = ["ClassInit"]
+    profilers = ["cl"]
+    jvmArgs = ["-Dio.perfmark.PerfMark.debug=true"]
+
+    /*
+    profilers = ["perfasm"]
+
+    jvmArgs = [
+            "-XX:+UnlockDiagnosticVMOptions",
+            "-XX:+LogCompilation",
+            "-XX:LogFile=/tmp/blah.txt",
+            "-XX:+PrintAssembly",
+            "-XX:+PrintInterpreter",
+            "-XX:+PrintNMethods",
+            "-XX:+PrintNativeNMethods",
+            "-XX:+PrintSignatureHandlers",
+            "-XX:+PrintAdapterHandlers",
+            "-XX:+PrintStubCode",
+            "-XX:+PrintCompilation",
+            "-XX:+PrintInlining",
+            "-XX:+TraceClassLoading",
+            "-XX:PrintAssemblyOptions=syntax",
+            "-XX:PrintAssemblyOptions=intel"
+    ]
+     */
+
+    //duplicateClassesStrategy DuplicatesStrategy.INCLUDE
+}