--- a/impl/src/main/java/io/perfmark/impl/SecretPerfMarkImpl.java
+++ b/impl/src/main/java/io/perfmark/impl/SecretPerfMarkImpl.java
@@ -387,4 +387,6 @@ final class SecretPerfMarkImpl {
       return ((gen >>> Generator.GEN_OFFSET) & 0x1L) != 0L;
     }
   }
+
+  private SecretPerfMarkImpl() {}
 }