--- a/impl/src/main/java/io/perfmark/impl/Storage.java
+++ b/impl/src/main/java/io/perfmark/impl/Storage.java
@@ -261,7 +261,5 @@ public final class Storage {
     }
   }
 
-  private Storage() {
-    throw new AssertionError("nope");
-  }
+  private Storage() {}
 }