```
//Before
Benchmark                                                               Mode  Cnt      Score     Error        Units
SecretPerfMarkImplClassInitBenchmark.forName_init                         ss  400  17209.431 ± 354.263        us/op
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.load             ss  400      3.599 ±   0.062  classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.load.norm        ss  400    152.523 ±   0.163   classes/op
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.unload           ss  400        ≈ 0            classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.unload.norm      ss  400        ≈ 0             classes/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit                       ss  400   6958.959 ± 133.175        us/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.load           ss  400      1.032 ±   0.018  classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.load.norm      ss  400     77.203 ±   0.224   classes/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.unload         ss  400        ≈ 0            classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400        ≈ 0             classes/op

//After, io.perfmark.PerfMark.debug false
Benchmark                                                               Mode  Cnt      Score     Error        Units
SecretPerfMarkImplClassInitBenchmark.forName_init                         ss  400  10837.033 ± 186.085        us/op
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.load             ss  400      1.407 ±   0.022  classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.load.norm        ss  400     83.015 ±   0.252   classes/op
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.unload           ss  400        ≈ 0            classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.unload.norm      ss  400        ≈ 0             classes/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit                       ss  400   6833.084 ± 100.928        us/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.load           ss  400      1.018 ±   0.016  classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.load.norm      ss  400     77.163 ±   0.230   classes/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.unload         ss  400        ≈ 0            classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400        ≈ 0             classes/op

//After, io.perfmark.PerfMark.debug true
Benchmark                                                               Mode  Cnt      Score     Error        Units
SecretPerfMarkImplClassInitBenchmark.forName_init                         ss  400  16939.197 ± 305.943        us/op
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.load             ss  400      3.553 ±   0.054  classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.load.norm        ss  400    152.603 ±   0.153   classes/op
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.unload           ss  400        ≈ 0            classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_init:·class.unload.norm      ss  400        ≈ 0             classes/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit                       ss  400   6975.875 ± 147.525        us/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.load           ss  400      1.034 ±   0.018  classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.load.norm      ss  400     77.313 ±   0.211   classes/op
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.unload         ss  400        ≈ 0            classes/sec
SecretPerfMarkImplClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400        ≈ 0             classes/op
```